+++ tests/Integration/MessagingTest.php
@@ -9,7 +9,9 @@ use Kreait\Firebase\Exception\Messaging\InvalidArgument;
 use Kreait\Firebase\Exception\Messaging\InvalidMessage;
 use Kreait\Firebase\Exception\Messaging\NotFound;
 use Kreait\Firebase\Exception\MessagingException;
+use Kreait\Firebase\Messaging\AndroidConfig;
 use Kreait\Firebase\Messaging\CloudMessage;
+use Kreait\Firebase\Messaging\MessageTarget;
 use Kreait\Firebase\Messaging\RawMessageFromArray;
 use Kreait\Firebase\Tests\IntegrationTestCase;
 
@@ -357,5 +359,43 @@ final class MessagingTest extends IntegrationTestCase
     {
         $this->expectException(NotFound::class);
         $this->messaging->getAppInstance(self::$unknownToken);
+    }
+
+    /**
+     * @see https://github.com/kreait/firebase-php/issues/713
+     */
+    public function testAndroidConfigTtlWorksWithAPositiveInt(): void
+    {
+        $config = AndroidConfig::fromArray([
+            'ttl' => 1,
+        ]);
+
+        $message = CloudMessage::withTarget(MessageTarget::TOKEN, $this->getTestRegistrationToken())
+            ->withAndroidConfig($config);
+
+        $this->messaging->send($message);
+
+        $this->addToAssertionCount(1);
+    }
+
+    /**
+     * @see https://github.com/kreait/firebase-php/issues/713
+     *
+     * @dataProvider \Kreait\Firebase\Tests\Unit\Messaging\AndroidConfigTest::validTtlValues
+     *
+     * @param int|string $value
+     */
+    public function testAndroidConfigTtlWorksWithValidValues($value): void
+    {
+        $config = AndroidConfig::fromArray([
+            'ttl' => $value,
+        ]);
+
+        $message = CloudMessage::withTarget(MessageTarget::TOKEN, $this->getTestRegistrationToken())
+            ->withAndroidConfig($config);
+
+        $this->messaging->send($message);
+
+        $this->addToAssertionCount(1);
     }
 }